9.1.20250216 release prep (#1890)
* 9.1.20250216 release prep

* updating gitignore and adding listeners

* adding default reportportal.properties

* fixing TestNGListener for CMD mode

* fixing strange null exception

Cannot invoke "com.shaft.properties.internal.Reporting.disableLogging()" because "com.shaft.driver.SHAFT$Properties.reporting" is null

* adding null check

* adding more null checks

* removing construction method

* tracing pesky null

* exposing pesky null

* finally caught it

* adding 10,000 null checks

+++ src/main/java/com/shaft/tools/io/internal/ReportManagerHelper.java
@@ -165,7 +165,8 @@ public class ReportManagerHelper {
 
     private static void initializeLogger() {
         // delete previous run execution log
-        FileActions.getInstance(true).deleteFile(System.getProperty("appender.file.fileName"));
+        if (System.getProperty("appender.file.fileName") != null)
+            FileActions.getInstance(true).deleteFile(System.getProperty("appender.file.fileName"));
         // initialize
         Configurator.initialize(null, PropertyFileManager.getCUSTOM_PROPERTIES_FOLDER_PATH() + "/log4j2.properties");
         logger = LogManager.getLogger(ReportManager.class.getName());
@@ -298,7 +299,7 @@ public class ReportManagerHelper {
     }
 
     public static void attachEngineLog(String executionEndTimestamp) {
-        if (!SHAFT.Properties.reporting.disableLogging()) {
+        if (SHAFT.Properties.reporting == null || !SHAFT.Properties.reporting.disableLogging()) {
             String engineLogCreated = "Successfully created attachment '" + SHAFT_ENGINE_LOGS_ATTACHMENT_TYPE + " - "
                     + "Execution log" + "'";
             var initialLoggingState = ReportManagerHelper.getDiscreteLogging();
@@ -430,7 +431,7 @@ public class ReportManagerHelper {
     }
 
     private static void createLogEntry(String logText, boolean addToConsoleLog) {
-        if (!SHAFT.Properties.reporting.disableLogging()) {
+        if (SHAFT.Properties.reporting == null || !SHAFT.Properties.reporting.disableLogging()) {
             String timestamp = (new SimpleDateFormat(TIMESTAMP_FORMAT)).format(new Date(System.currentTimeMillis()));
             if (logText == null) {
                 logText = "null";
@@ -508,7 +509,7 @@ public class ReportManagerHelper {
      * @param logText the text that needs to be logged in this action
      */
     public static void writeStepToReport(String logText) {
-        if (!SHAFT.Properties.reporting.disableLogging()) {
+        if (SHAFT.Properties.reporting == null || !SHAFT.Properties.reporting.disableLogging()) {
             createLogEntry(logText, true);
             Allure.step(logText, getStepStatus(logText));
         }
@@ -643,7 +644,7 @@ public class ReportManagerHelper {
     }
 
     public static void log(String logText, List<List<Object>> attachments) {
-        if (!SHAFT.Properties.reporting.disableLogging()) {
+        if (SHAFT.Properties.reporting == null || !SHAFT.Properties.reporting.disableLogging()) {
             if (!logText.toLowerCase().contains("failed") && getDiscreteLogging() && isInternalStep()) {
                 createLogEntry(logText, Level.INFO);
                 if (attachments != null && !attachments.isEmpty() && (attachments.size() > 1 || (attachments.getFirst() != null && !attachments.getFirst().isEmpty()))) {
@@ -678,7 +679,7 @@ public class ReportManagerHelper {
         CheckpointType type = (logText.toLowerCase().contains("verification")) ? CheckpointType.VERIFICATION : CheckpointType.ASSERTION;
 
         if (type.equals(CheckpointType.VERIFICATION) && status.equals(CheckpointStatus.FAIL)
-                || !SHAFT.Properties.reporting.disableLogging()) {
+                || SHAFT.Properties.reporting == null || !SHAFT.Properties.reporting.disableLogging()) {
             if (customLogMessages != null && !customLogMessages.isEmpty() && !customLogMessages.getFirst().trim().isEmpty()) {
                 String customLogText = customLogMessages.getFirst();
                 if (status == CheckpointStatus.PASS) {
